refact: abastract common refund stripe methods
